bot/requirements.txt
aiogram
python-telegram-bot
python-dotenv
groq
psycopg[binary]